requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile
#
aiohttp==3.8.3
    # via discord-py
aiosignal==1.3.1
    # via aiohttp
async-timeout==4.0.2
    # via aiohttp
attrs==22.1.0
    # via aiohttp
cachetools==5.2.0
    # via
    #   -r requirements.in
    #   google-auth
certifi==2022.12.7
    # via requests
charset-normalizer==2.1.1
    # via
    #   aiohttp
    #   requests
discord-py==2.1.0
    # via -r requirements.in
frozenlist==1.3.3
    # via
    #   aiohttp
    #   aiosignal
google-api-core==2.11.0
    # via google-api-python-client
google-api-python-client==2.70.0
    # via -r requirements.in
google-auth==2.15.0
    # via
    #   google-api-core
    #   google-api-python-client
    #   google-auth-httplib2
google-auth-httplib2==0.1.0
    # via google-api-python-client
googleapis-common-protos==1.57.0
    # via google-api-core
httplib2==0.21.0
    # via
    #   google-api-python-client
    #   google-auth-httplib2
idna==3.4
    # via
    #   requests
    #   yarl
multidict==6.0.3
    # via
    #   aiohttp
    #   yarl
protobuf==4.21.12
    # via
    #   google-api-core
    #   googleapis-common-protos
pyasn1==0.4.8
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.2.8
    # via google-auth
pyparsing==3.0.9
    # via httplib2
requests==2.28.1
    # via google-api-core
rsa==4.9
    # via google-auth
six==1.16.0
    # via
    #   google-auth
    #   google-auth-httplib2
uritemplate==4.1.1
    # via google-api-python-client
urllib3==1.26.13
    # via requests
yarl==1.8.2
    # via aiohttp
